requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
idna==3.7
PyYAML==6.0.1
requests==2.32.3
tqdm==4.66.4
urllib3==2.2.2
